requirements.txt
nemo_toolkit[asr]
fastapi
uvicorn
torchaudio
python-multipart